refactor: move util package to global
