+++ src/main/java/com/dnd/wedding/domain/oauth/handler/OAuth2AuthenticationFailureHandler.java
@@ -3,7 +3,7 @@ package com.dnd.wedding.domain.oauth.handler;
 import static com.dnd.wedding.domain.jwt.repository.CookieAuthorizationRequestRepository.REDIRECT_URI_PARAM_COOKIE_NAME;
 
 import com.dnd.wedding.domain.jwt.repository.CookieAuthorizationRequestRepository;
-import com.dnd.wedding.global.config.util.CookieUtil;
+import com.dnd.wedding.global.util.CookieUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
