+++ src/main/java/com/dnd/wedding/domain/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -5,8 +5,8 @@ import static com.dnd.wedding.domain.jwt.repository.CookieAuthorizationRequestRe
 
 import com.dnd.wedding.domain.jwt.JwtTokenProvider;
 import com.dnd.wedding.domain.jwt.repository.CookieAuthorizationRequestRepository;
-import com.dnd.wedding.global.config.util.CookieUtil;
 import com.dnd.wedding.global.exception.BadRequestException;
+import com.dnd.wedding.global.util.CookieUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
